+++ src/components/route/AppRoutes.tsx
@@ -10,7 +10,6 @@ import CaseListPage from "../case/CaseListPage";
 import CasesPage from "../case/CasesPage";
 import ClientDetailPage from "../client/ClientDetailPage";
 import ClientsPage from "../client/ClientsPage";
-import EmployeePage from "../employee/EmployeePage";
 import NotFoundPage from "../error/NotFoundPage";
 import HomePage from "../home/HomePage";
 import JoinPage from "../join/JoinPage";
@@ -18,10 +17,9 @@ import Layout from "../layout/Layout";
 import LoginPage from "../login/LoginPage";
 import EmployeeLayout from "../employee/EmployeeLayout";
 import PrivatePage from "../private/PrivatePage";
-import employeeButtonIdState from "../../states/employee/EmployeeButtonIdState";
 import EmployeeListPage from "../employee/list/EmployeesListPage";
 import employeeIdState from "../../states/employee/EmployeeIdState";
-import EmployeeCasePage from "../employee/case/list/EmployeeCasePage";
+import EmployeeDetailPage from "../employee/case/EmployeeDetailPage";
 import caseTabIdState from "../../states/case/CaseTabIdState.tsx";
 import ValidatePage from "../join/ValidatePage";
 import SchedulePage from "../schedule/SchedulePage.tsx";
@@ -35,7 +33,6 @@ import BalanceIcon from "@mui/icons-material/Balance";
 import { MemberInfo } from "../employee/type/MemberInfo.tsx";
 import roleListState from "../../states/data/roleListState.ts";
 import snbLoadedState from "../../states/common/SnbLoadedState.ts";
-import EmployeeDetailPage from "../employee/detail/EmployeeDetailPage.tsx";
 import privateButtonIsClickState from "../../states/private/PrivateButtonIsClickState";
 
 function AppRoutes() {
@@ -45,20 +42,20 @@ function AppRoutes() {
   const [mainNavigationBar, setMainNavigationBar] = useRecoilState(
     mainNavigationBarState,
   );
-  const setSubNavigationBar = useSetRecoilState(subNavigationBarState);
+  const [subNavigationBar, setSubNavigationBar] = useRecoilState(
+    subNavigationBarState,
+  );
   const [clientId, setClientId] = useRecoilState(clientIdState);
   const setCaseId = useSetRecoilState(caseIdState);
   const setEmployeeId = useSetRecoilState(employeeIdState);
   const setSnbLoaded = useSetRecoilState(snbLoadedState);
 
-  const setEmployeeButtonId = useSetRecoilState(employeeButtonIdState);
   const setCaseTabId = useSetRecoilState(caseTabIdState);
   const setScheduleButtonIsClick = useSetRecoilState(
     scheduleButtonIsClickState,
   );
   const setPrivateButtonIsClick = useSetRecoilState(privateButtonIsClickState);
   const roleList = useRecoilValue(roleListState);
-  const employeeButton = useRecoilValue(employeeButtonIdState);
 
   useEffect(() => {
     const { pathname, search } = location;
@@ -309,26 +306,6 @@ function AppRoutes() {
       return;
     }
 
-    // /employees/list
-    if (length === 2 && paths[1] === "employees" && paths[2] === "list") {
-      setMainNavigationBar({
-        ...mainNavigationBar,
-        curId: 2,
-      });
-      setSubNavigationBar({
-        type: "none",
-        curId: -1,
-        items: [],
-      });
-      setSnbLoaded(true);
-      setClientId(null);
-      setCaseId(null);
-      setEmployeeId(null);
-
-      setEmployeeButtonId(1);
-      return;
-    }
-
     // /employees/:employeeId
     if (
       length === 2 &&
@@ -336,6 +313,13 @@ function AppRoutes() {
       paths[2] &&
       !isNaN(Number(paths[2]))
     ) {
+      if (subNavigationBar.type === "employee") {
+        setClientId(null);
+        setCaseId(null);
+        setEmployeeId(Number.parseInt(paths[2]));
+        return;
+      }
+
       setMainNavigationBar({
         ...mainNavigationBar,
         curId: 2,
@@ -351,10 +335,7 @@ function AppRoutes() {
                 text: item.name,
                 subText: roleList.filter((it) => it.id == item.roleId)[0]
                   .nameKr,
-                url:
-                  employeeButton === 2
-                    ? `employees/${item.id}`
-                    : `employees/${item.id}/cases`,
+                url: `employees/${item.id}`,
                 SvgIcon: BalanceIcon,
               };
             },
@@ -371,54 +352,6 @@ function AppRoutes() {
       setCaseId(null);
       setEmployeeId(Number.parseInt(paths[2]));
 
-      setEmployeeButtonId(2);
-      return;
-    }
-
-    // /employees/:employeeId/cases
-    if (
-      length === 3 &&
-      paths[1] === "employees" &&
-      paths[2] &&
-      !isNaN(Number(paths[2])) &&
-      paths[3] === "cases"
-    ) {
-      setMainNavigationBar({
-        ...mainNavigationBar,
-        curId: 2,
-      });
-      setSnbLoaded(false);
-      requestDeprecated("GET", `/members/employees`, {
-        onSuccess: (res) => {
-          const memberInfos: MemberInfo[] = res.data.memberDtoNonPassList;
-          const newItems: SubNavigationBarItemState[] = memberInfos.map(
-            (item) => {
-              return {
-                id: item.id,
-                text: item.name,
-                subText: roleList.filter((it) => it.id == item.roleId)[0]
-                  .nameKr,
-                url:
-                  employeeButton === 2
-                    ? `employees/${item.id}`
-                    : `employees/${item.id}/cases`,
-                SvgIcon: BalanceIcon,
-              };
-            },
-          );
-          setSubNavigationBar({
-            type: "employee",
-            curId: newItems[0].id,
-            items: newItems,
-          });
-          setSnbLoaded(true);
-        },
-      });
-      setClientId(null);
-      setCaseId(null);
-      setEmployeeId(Number.parseInt(paths[2]));
-
-      setEmployeeButtonId(3);
       return;
     }
 
@@ -582,13 +515,10 @@ function AppRoutes() {
           <Route path=":caseId" element={<CaseDetailPage />} />
         </Route>
         <Route path="employees" element={<EmployeeLayout />}>
-          {/* /employees */}
-          <Route index element={<EmployeePage />} />
-          {/*/employees/list*/}
-          <Route path={"list"} element={<EmployeeListPage />} />
+          {/* /employees/}*/}
+          <Route path={""} element={<EmployeeListPage />} />
           {/* /employees/:employeeId */}
           <Route path=":employeeId" element={<EmployeeDetailPage />} />
-          <Route path=":employeeId/cases" element={<EmployeeCasePage />} />
         </Route>
         <Route path="schedule" element={<SchedulePage />} />
         <Route path="private" element={<PrivatePage />} />
