--- src/states/employee/EmployeeButtonIdState.tsx
@@ -1,8 +1,0 @@
-import { atom } from "recoil";
-
-const employeeButtonIdState = atom({
-  key: "employeeButtonIdState",
-  default: -1,
-});
-
-export default employeeButtonIdState;
